Better info in the calendar events
Code clean-up because `weekday()` is 0-based. Code more robust as well
because arrays have been defined as constants. Because you can push
data into an array, even when it's a constant. Some refactoring.

Also redefined how the `places.js` module is exposed. The `list()`
method as been removed which should make the code a bit easier to read.