Use calendar and switch on light
When this script is run (i.e. via a cron job or something similar), it
will look up the reminder calendar that's hosted on Github and check
whether there is a reminder in it that matches today's date.  If so,
it will select the right colour for the rubbish in question and switch
on the Philips Hue light that's identified by the LIGHT_URL.